@@ -4,7 +4,7 @@ version: 1.5.0
 homepage: https://github.com/boeledi/RangeSlider
 
 environment:
-    sdk: ">=1.19.0 <3.0.0"
+    sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
